New HRI framework/new_framework_microservices/emotion_service/requirements.txt
Flask
tensorflow==1.14.0
keras==2.3.1
statistics
pandas
matplotlib==2.2.4
pillow
python-logstash
gunicorn==19.10.0
py_eureka_client==0.7.4
python-json-logger
